refactor(frontend): extract resize math to shared module
Move snap, snapToEdges, applyEdgeResize, applyCornerResize and their
constants from eink-dashboard-card.ts to a new resize-math.ts module.
The card imports and re-exports them for backward compatibility.
Tests move to test/resize-math.test.ts.
